+++ index.js
@@ -1,7 +1,7 @@
-const server = require('./server.js');
-const setupCRI = require('./debugging-interface.js');
-const setupChromeLauncher = require('./chrome-launcher.js');
-const setupTemplateRenderer = require('./template-renderer.js');
+const server = require('./src/main/server');
+const setupCRI = require('./src/main/debugging-interface.js');
+const setupChromeLauncher = require('./src/main/chrome-launcher.js');
+const setupTemplateRenderer = require('./src/main/template-renderer.js');
 
 setupChromeLauncher();
 server.startServer();
