+++ src/main/chrome-launcher.js
