+++ src/main/debugging-interface.js
