+++ src/main/template-renderer.js
